# api/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "operator_pool.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "OperatorPoolService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/operator/list": {
      "get": {
        "operationId": "OperatorPoolService_ListOperators",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/operator_pool_serviceListOperatorsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "OperatorPoolService"
        ]
      }
    },
    "/operator/next": {
      "get": {
        "operationId": "OperatorPoolService_GetNext",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/operator_pool_serviceGetNextResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "OperatorPoolService"
        ]
      }
    },
    "/operator/stats": {
      "get": {
        "operationId": "OperatorPoolService_GetStats",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/operator_pool_serviceGetStatsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "OperatorPoolService"
        ]
      }
    },
    "/operator/status": {
      "post": {
        "operationId": "OperatorPoolService_SetStatus",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/operator_pool_serviceSetStatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/operator_pool_serviceSetStatusRequest"
            }
          }
        ],
        "tags": [
          "OperatorPoolService"
        ]
      }
    }
  },
  "definitions": {
    "operator_pool_serviceGetNextResponse": {
      "type": "object",
      "properties": {
        "operatorId": {
          "type": "string"
        }
      }
    },
    "operator_pool_serviceGetStatsResponse": {
      "type": "object",
      "properties": {
        "available": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "operator_pool_serviceListOperatorsResponse": {
      "type": "object",
      "properties": {
        "operators": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/operator_pool_serviceOperatorStatus"
          }
        }
      }
    },
    "operator_pool_serviceOperatorStatus": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "available": {
          "type": "boolean"
        },
        "activeSessions": {
          "type": "integer",
          "format": "int32"
        },
        "maxSessions": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "operator_pool_serviceSetStatusRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "available": {
          "type": "boolean"
        },
        "maxSessions": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "operator_pool_serviceSetStatusResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}